requirements.txt
dnspython
graphviz
ipaddress
